+++ src/include/util/graph.h
@@ -18,7 +18,7 @@ class Graph
 	private:
 		struct SchedulePoint
 		{
-			float coord;
+			double coord;
 			unsigned index;
 			bool start;
 		};
